# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/workouts/workout-report/": {
            "post": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Create a WorkoutReport",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "WorkoutReport"
                ],
                "summary": "Create a WorkoutReport",
                "parameters": [
                    {
                        "description": "Create a WorkoutReport",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.CreateWorkoutReportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "WorkoutReport response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.WorkoutReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/workouts/workout-report/{id}": {
            "get": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Get a WorkoutReport by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkoutReport"
                ],
                "summary": "Get a WorkoutReport by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WorkoutReport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "WorkoutReport response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.WorkoutReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Update a WorkoutReport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkoutReport"
                ],
                "summary": "Update a WorkoutReport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WorkoutReport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update a WorkoutReport",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.UpdateWorkoutReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "WorkoutReport response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.WorkoutReportResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AuthBearer": []
                    }
                ],
                "description": "Delete a WorkoutReport",
                "tags": [
                    "WorkoutReport"
                ],
                "summary": "Delete a WorkoutReport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WorkoutReport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.CreateWorkoutReportRequest": {
            "type": "object",
            "required": [
                "details",
                "workout_id"
            ],
            "properties": {
                "details": {
                    "type": "string"
                },
                "workout_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.UpdateWorkoutReportRequest": {
            "type": "object",
            "required": [
                "details",
                "workout_id"
            ],
            "properties": {
                "details": {
                    "type": "string"
                },
                "workout_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alielmi98_go-hexa-workout_internal_workout_adapter_http_dto.WorkoutReportResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "workout_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alielmi98_go-hexa-workout_pkg_helper.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "result": {},
                "resultCode": {
                    "$ref": "#/definitions/github_com_alielmi98_go-hexa-workout_pkg_helper.ResultCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_alielmi98_go-hexa-workout_pkg_helper.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                40001,
                40101,
                40301,
                40401,
                42901,
                42902,
                50001,
                50002,
                50003,
                50004,
                50005,
                40002
            ],
            "x-enum-varnames": [
                "Success",
                "ValidationError",
                "AuthError",
                "ForbiddenError",
                "NotFoundError",
                "LimiterError",
                "OtpLimiterError",
                "CustomRecovery",
                "InternalError",
                "InvalidInputError",
                "DatabaseError",
                "UnknownError",
                "BadRequest"
            ]
        }
    },
    "securityDefinitions": {
        "AuthBearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}